backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
reportlab==4.0.7
jinja2==3.1.2
python-multipart==0.0.6
streamlit==1.28.0
streamlit-sortables==0.1.0
python-dotenv==1.0.0
